Add migration files

+++ database/migrations/2019_01_13_144534_create_property_table.php
@@ -17,13 +17,14 @@ class CreatePropertyTable extends Migration
             $table->increments('id');
             $table->string('name', 100)->unique();
             $table->string('description', 255)->nullable();
-            $table->tinyInteger('property_type_id');
+            $table->unsignedInteger('property_type_id');
             $table->string('gps', 50)->nullable();
             $table->tinyInteger('is_active')->default(1);
             $table->text('custom_fields')->nullable();
             $table->integer('created_by')->nullable();
             $table->integer('updated_by')->nullable();
             $table->timestamps();
+            // $table->foreign('property_type_id')->references('id')->on('tbl_properties_types')->onUpdate('cascade')->onDelete('cascade');
         });
     }
